rename 'Running Steps' to 'Workflow Steps'
as the action will display all steps that have both run and are running to
call the action "Running steps" is incorrect and at first glance to a user
is incorrect. This changes the name to "Workflow Steps" which is
marginally better.

Originally-Committed-As: a5c41df9d83e717f337da74a5b3aa04b4057e088