Reworked to an OptionalJobProperty to disable concurrent builds.
Having an OptionalJobProperty used solely as a boolean, effectively,
means we need to have the default case be in effect when the property
doesn't exist. That's different than the traditional concurrent build
check, where the default is to not allow concurrent builds. Since we
do want concurrent builds by default in Pipeline jobs, we need to
instead have the property be for *disabling* concurrent
builds. So...yeah.

Not sure yet how to do a UI config test for an OptionalJobProperty -
need to work on that.