So this generally works now.
That said, it needs tweaks - the UI needs work, since currently it
just shows up in the job properties, with no break between the
triggers and subsequent properties, which is not a good
experience. And while I've verified by hand that triggers are
successfully stopped, started, persisted through a config round trip,
etc, actual tests are still very much needed.

That said, this *does* generally work!