[FIXED JENKINS-38454] Make sure we add the triggers before starting.
This was...complex. But it works. We make sure the triggers property
has been added before starting, in case the start method for a trigger
class expects the trigger to already be present on the job. Existing
tests all pass, and the new test failed without the changes to
WorkflowJob and PipelineTriggersJobProperty. Phew.

Note - I'm working on getting rid of the need for posting hardcoded
JSON and just using form manipulation.