Remove unnecessary stops/starts of triggers
In the case where you are using the properties() step with job properties other that the PipelineTriggersJobProperty, we do not want
stops and starts of a trigger everytime we update a non-PipelineTriggersJobProperty.

This can pose a problem for a trigger that may have established a connection to an external system and it is forced to stop/start
frequently.