WorkflowRun: Allow retrieval of logCopyGuard without lock.
`WorkflowRun::logCopyGuard` is a private transient object that's used
to ensure mutual exclusion of certain operations. Unfortunately, the
getter for it (which initializes the field if it is `null`, as it
would be after deserialization) was `synchronized`, and thus could not
run at the same time as other potentially expensive synchronized
methods (like `getExecution`, which loads a `build.xml`, or `save`,
which writes one).

To work around, we implement a `readObject` method that will
explicitly initialize the `logCopyGuard` object. Now, the
`getLogCopyGuard` method can assume it is non-`null`, and therefore
does not need to be `synchronized`.